feat: Enhance search functionality
The search feature in the ultra-market app has been significantly improved.

The API route now supports searching not only by item name, but also by description and tags. This change allows for more comprehensive and accurate results when users perform a search.

On the front-end, the initial state of the search field is now set to decodeURI(params.search?.[0]) instead of params.search?.[0]. This ensures that any special characters in the URL are correctly interpreted, providing a better user experience.